+++ src/app/_components/local-status-box/searchHistoryBox.tsx
