+++ src/app/_components/navBar.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import React from "react";
 import Link from "next/link";
-import { ModeToggle } from "@/app/_components/modeToggle";
+import { ModeToggle } from "@/components/ui/modeToggle";
 
 export default function NavBar() {
   return (
